requirements.txt
pyperclip
uuid
requests
pycryptodome
pywin32